requirements.txt
requests
yfinance
pandas
python-dotenv